Add read only open flag
gsd_open now accpets a flag that specifies if the file is to be opened read-only or read-write. It passes this flag down to the open() syscall and also saves the flag for later. API calls that would write to the file check if the file is in a read only mode and error out immediately.