Add in error checking if fl.pyx is updated
When fl.pyx is updated, fl?.c are out of date. Add SHA256 based error
checking to determine if this is the case. It will present users without
cython with a reaonsable error message - only when they try building a
commit that doesn't have an up to date code generation.