Expand index entry size to 32 bytes
Fixes an error in the specification. The previous 28-byte index
structure was padded to 32 bytes due to alignment constraints. Structure
alignment is not guarunteed in the C standard, but all current compilers
doe sane things, so we are mostly safe. If we discover compilers that do
not work with this struct layout, we can add appropriate pragmas or
attributes as needed.