Implement _HOOMDTrajectoryView class.
This patch modifies the HOOMDTrajectory class to return a view after
slicing instead of an iterable. A view represents a subset of a
trajectory that can still be queried for length, iterated over or, and
that is the key point of this, used for invididual frame selection or
sliced again.

For example:

  with gsd.hoomd.open('trajectory.gsd') as traj:
      # Select subset (every 10th frame):
      view = traj[::10]

      # This was already possible prior to this patch:
      for frame in view:
          pass

      # But now we can also select frames from a view:
      n_th_frame = view[n]

      # And slice it again:
      for frame in view[:10]:
          pass

A sliced trajectory view returns a new view of the trajectory, similar
to a numpy array view.