Remove wrong install_requires on cython
Cython is used only at build time, and the resulting Python extensions
do not need Cython installed.  Therefore, it should not be listed
in install_requires.

Furthermore, the sdist archives include the generated .c files that also
do not need Cython installed to build.  For this reason, Cython is not
listed in setup_requires either.