Use fcntl(.., F_FULLSYNC)
Apparently, fsync on mac does little to nothing. That is why the Mac
benchmarks were so much better than everything else. fcntl(fd,
F_FULLSYNC) is the correct way to sync on Mac, however it seems much
slower than it should be (or perhaps Apple's SSDs are not as fast as
they should be). This change will drastically reduce performance for Mac
users. However, the data files will be must less prone to corruption as
a result.

Here are several discussions online about fsync vs fcntl(.., F_FULLSYNC)
https://mjtsai.com/blog/2022/02/17/apple-ssd-benchmarks-and-f_fullsync/
https://stackoverflow.com/questions/78626405/surprisingly-high-latency-for-rusts-sync-all-on-macos
https://news.ycombinator.com/item?id=30372218
https://bonsaidb.io/blog/acid-on-apple/

Most of these posts mention the performance loss on M1 or newer
hardware. It is not clear if intel macs have the same problem.

GSD files would be relatively corruption resistent with a BARRIERSYNC,
however there is no cross platform way to issue a barrier sync. Ideally,
GSD will provide the same corruption safety on all platforms.