[breaking] Do not automatically flush the index buffer.
Give the caller complete control over when the index buffer is comitted
to disk. The write buffer now exists solely to buffer writes of small
data chunks and it will continue to be automatically flushed. Decoupling
the flush of the index buffer from the data buffer drastically increases
write throughput on low latency file systems. It allows the OS to buffer
the streaming data writes as appropriate without constant fsync
interruptions. In most GSD files, the majority of the file is taken up
by data chunks.